% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{ffplot}
\alias{ffplot}
\title{Create a plot of the fractal}
\usage{
ffplot(df, x = x, y = y)
}
\arguments{
\item{df}{tibble with the data}

\item{x}{x-coordinates}

\item{y}{y-coordinates}
}
\value{
a ggplot object
}
\description{
Create a plot of the fractal
}
