% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hilbert.R
\name{flip_uperright_lowerleft}
\alias{flip_uperright_lowerleft}
\title{Flip the data along the diagonal between the left-upper and right-lower corner}
\usage{
flip_uperright_lowerleft(df, n)
}
\arguments{
\item{df}{tibble with the columns x and y}

\item{n}{number of iterations}
}
\value{
The flipped tibble
}
\description{
Flip the data along the diagonal between the left-upper and right-lower corner
}
