% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hilbert.R
\name{flip_x_y}
\alias{flip_x_y}
\title{Flip the data along the diagonal between the left-lower and right-upper corner}
\usage{
flip_x_y(df)
}
\arguments{
\item{df}{tibble with the columns x and y}
}
\value{
The flipped tibble
}
\description{
Flip the data along the diagonal between the left-lower and right-upper corner
}
