% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hilbert.R
\name{hilbert_curve}
\alias{hilbert_curve}
\title{Calculation of the Hilbert curve}
\usage{
hilbert_curve(n, prev = tibble(x = 0, y = 0))
}
\arguments{
\item{n}{Iteration}

\item{prev}{tibble of the previous iteration}
}
\value{
The tibble with the x and y coordinates for the Hilbert curve
}
\description{
Recursive function that calculates the Hilbert curve after n iterations.
}
