% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hilbert.R
\name{move_curve}
\alias{move_curve}
\title{Move the data so that the start point of the curve is changed}
\usage{
move_curve(df, new_start)
}
\arguments{
\item{df}{tibble with x and y coordinates}

\item{new_start}{tibble with one x and y coordinate as new starting point}
}
\value{
the moved tibble
}
\description{
Move the data so that the start point of the curve is changed
}
